Mise à jour demandée
- J'ai imaginé la solution suivante pour la création de channel dynamique : un tableau dans le fichier config qui permet de facilement ajouter un ID qui rendrait un channel vocal "dynamique", le comportement est toujours le même, on clone le channel. Seul soucis c'est qu'on risque de s'y perdre dans le fichiers config. Plusieurs solution pour ça : repenser la structure, commeter le fichier config.
- J'ai ajouté la permission "MANAGE_CHANNEL" au créateur du channel. Ajouter MANAGE_ROLES lui permet d'avoir beaucoup plus de controle sur le channel mais ça créer aussi des effets bizare (discord lui dit qu'il peut cloner et créer des channels mais ça crash quand il clique sur OK)
- J'ai nettoyé le code voiceStateUpdate (supprimer quoi x) )

+++ events/voiceChannelLeave.js
@@ -1,10 +1,6 @@
-const config = require("../config.js");
-
 module.exports = async (client, logs, member, channel) => {
-    if (client.colearningChannels.get(channel.id)) {
-        if (channel.members.size <= 0) {
-            channel.delete();
-            logs.info("There was no members left in that co-learning channel, it was deleted");
-        }
+    if (client.dynamicVoiceChannels.has(channel.id) && channel.members.size === 0) {
+        channel.delete();
+        logs.info(`There was no members left in a dynamic voice channel , it was deleted`);
     }
 };
